// src/app/professionals/models/types.ts
import type { UseQueryStatesKeysMap } from "nuqs";

import type { FiltersWithPagination } from "@/lib/types/global-types";

export type ProfessionalFilters = UseQueryStatesKeysMap<FiltersWithPagination<{
  profession_id: number;
  last_names: string;
  names: string;
  email: string;
  rut: string;
  id: number;
}>>;